Move servlet-specific functionality into its own maven module
This also moves some of the generic code in the servlet
transports/connections into the common module.
