+++ socket-io-servlet/pom.xml
@@ -6,7 +6,7 @@
     <parent>
         <artifactId>socketio-parent</artifactId>
         <groupId>com.codeminders.socketio</groupId>
-        <version>1.0.9</version>
+        <version>2.0.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>socket-io-servlet</artifactId>
@@ -19,7 +19,7 @@
         <dependency>
             <groupId>com.codeminders.socketio</groupId>
             <artifactId>socket-io</artifactId>
-            <version>${project.version}</version>
+            <version>2.0.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>javax.servlet</groupId>
