+++ socket-io-servlet/src/main/java/com/codeminders/socketio/server/servlet/transport/HttpServletRequestWrapper.java
@@ -1,4 +1,4 @@
-package com.codeminders.socketio.server.transport;
+package com.codeminders.socketio.server.servlet.transport;
 
 import com.codeminders.socketio.server.HttpRequest;
