+++ socket-io-servlet/src/main/java/com/codeminders/socketio/server/servlet/transport/JSONPPollingTransport.java
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  *
  */
-package com.codeminders.socketio.server.transport;
+package com.codeminders.socketio.server.servlet.transport;
 
 import com.codeminders.socketio.protocol.EngineIOProtocol;
 import com.codeminders.socketio.server.HttpRequest;
