+++ socket-io-servlet/src/main/java/com/codeminders/socketio/server/servlet/transport/websocket/SynchronizedWebsocketIO.java
@@ -1,19 +1,17 @@
 /**
  * The MIT License
- * Copyright (c) 2010 Tad Glines
- *
- * Contributors: Ovea.com, Mycila.com
- *
+ * Copyright (c) 2018 Alex Saveliev (lyolik@codeminders.com)
+ * <p/>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p/>
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- *
+ * <p/>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -22,30 +20,28 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.codeminders.socketio.server.transport;
+package com.codeminders.socketio.server.servlet.transport.websocket;
 
-import com.codeminders.socketio.server.TransportConnection;
-import com.codeminders.socketio.server.TransportType;
+import java.io.IOException;
 
-import javax.servlet.ServletConfig;
-import javax.servlet.ServletContext;
+/**
+ * @author Alex Saveliev (lyolik@codeminders.com)
+ */
+public class SynchronizedWebsocketIO extends WebsocketIO {
 
-public class XHRPollingTransport extends AbstractHttpTransport
-{
-    public XHRPollingTransport(ServletConfig servletConfig, ServletContext servletContext) {
-        super(servletConfig, servletContext);
+    public SynchronizedWebsocketIO(javax.websocket.Session remoteEndpoint) {
+        super(remoteEndpoint);
     }
 
-    @Override
-    public TransportType getType()
-    {
-        return TransportType.XHR_POLLING;
+    public synchronized void sendString(String data) throws IOException {
+        super.sendString(data);
     }
 
-    @Override
-    public TransportConnection createConnection()
-    {
-        return new XHRTransportConnection(this);
+    public synchronized void sendBinary(byte[] data) throws IOException {
+        super.sendBinary(data);
     }
 
+    public void disconnect() throws IOException {
+        remoteEndpoint.close();
+    }
 }
